perms

+++ backend/rest_api/views/common.py
@@ -14,8 +14,11 @@ logger = logging.getLogger("app")
 
 
 class UserLogoutAPI(generics.GenericAPIView):
+    permission_classes = [permissions.AllowAny]
+
     def post(self, request, *args, **kwargs):
-        AuthToken.objects.filter(user=request.user).delete()
+        if request.user.is_authenticated:
+            AuthToken.objects.filter(user=request.user).delete()
         return Response({'message': 'Logged out successfully'}, status=status.HTTP_200_OK)
 
 
@@ -45,7 +48,6 @@ class UserLoginAPI(generics.GenericAPIView):
         user = serializer.validated_data
         user.save()
         AuthToken.objects.filter(user=user).delete()
-        print("here")
         logger.info("Logged in %s", user.get_name())
 
         response_data = {
